+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.1.0-rc.1] - 2020-04-04
+> **Added**
+>> - New Tab: Restore to the previous version
+>> - versions.txt file to store the current and the backup versions (if backup available)
+
 ## [1.0.0-rc.1] - 2020-04-02
 > **Added**
 >> - Download online version